refactor(backend): split search/services.py by class
Decomposes the 513-LOC search/services.py into a sub-package, one
file per class:

  search/services/
    __init__.py               (31)   re-exports all 4 classes
    pagination.py             (22)   PaginationParams dataclass
    global_search.py          (88)   GlobalSearchService
    phenopacket_search.py     (203)  PhenopacketSearchService
    facet.py                  (206)  FacetService

Every file is under 500 LOC. Public import path unchanged:

    from app.search.services import (
        PaginationParams,
        GlobalSearchService,
        PhenopacketSearchService,
        FacetService,
    )

tests/test_global_search.py imports all four symbols this way and
continues to work unchanged.

Counts unchanged: 907 passed, 9 skipped, 3 xfailed.
